aws-bilder

+++ src/main/java/sandrakorpi/molnintegrationbookapp/DTOs/BookDto.java
@@ -7,23 +7,14 @@ package sandrakorpi.molnintegrationbookapp.DTOs;
  */
 public class BookDto {
 
-    private int bookId;
+    private long bookId;
     private String title;
     private String author;
     private int yearPublished;
     private String genre;
     private boolean recommended;
 
-    /**
-     * Konstruktor för att skapa en ny bok DTO.
-     *
-     * @param bookId ID för boken
-     * @param title Titel på boken
-     * @param author Författare till boken
-     * @param yearPublished År boken publicerades
-     * @param genre Genre av boken
-     * @param isRecommended Om boken är rekommenderad eller inte
-     */
+
     public BookDto(int bookId, String title, String author, int yearPublished, String genre, boolean isRecommended) {
         this.bookId = bookId;
         this.title = title;
@@ -37,110 +28,51 @@ public class BookDto {
     public BookDto() {
     }
 
-    /**
-     * Hämtar bok-ID.
-     *
-     * @return bok-ID
-     */
-    public int getBookId() {
+    public long getBookId() {
         return bookId;
     }
-
-    /**
-     * Sätter bok-ID.
-     *
-     * @param bookId bok-ID att sätta
-     */
-    public void setBookId(int bookId) {
+    public void setBookId(long bookId) {
         this.bookId = bookId;
     }
 
-    /**
-     * Hämtar titel på boken.
-     *
-     * @return titel på boken
-     */
     public String getTitle() {
         return title;
     }
 
-    /**
-     * Sätter titel på boken.
-     *
-     * @param title titel att sätta
-     */
     public void setTitle(String title) {
         this.title = title;
     }
 
-    /**
-     * Hämtar författare av boken.
-     *
-     * @return författare av boken
-     */
     public String getAuthor() {
         return author;
     }
 
-    /**
-     * Sätter författare av boken.
-     *
-     * @param author författare att sätta
-     */
     public void setAuthor(String author) {
         this.author = author;
     }
 
-    /**
-     * Hämtar året boken publicerades.
-     *
-     * @return år boken publicerades
-     */
     public int getYearPublished() {
         return yearPublished;
     }
 
-    /**
-     * Sätter året boken publicerades.
-     *
-     * @param yearPublished år att sätta
-     */
     public void setYearPublished(int yearPublished) {
         this.yearPublished = yearPublished;
     }
 
-    /**
-     * Hämtar genre av boken.
-     *
-     * @return genre av boken
-     */
+
     public String getGenre() {
         return genre;
     }
 
-    /**
-     * Sätter genre av boken.
-     *
-     * @param genre genre att sätta
-     */
+
     public void setGenre(String genre) {
         this.genre = genre;
     }
 
-    /**
-     * Kontrollerar om boken är rekommenderad.
-     *
-     * @return true om boken är rekommenderad, annars false
-     */
     public boolean isRecommended() {
         return recommended;
     }
 
-    /**
-     * Sätter om boken är rekommenderad.
-     *
-     * @param isRecommended true om boken ska vara rekommenderad, annars false
-     */
     public void setIsRecommended(boolean isRecommended) {
         this.recommended = isRecommended;
     }
